requirements.txt
blinker==1.8.2
click==8.1.7
Faker==26.0.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
